news_sentiment_trader/requirements.txt
transformers==4.40.0
torch>=2.1.0
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
yfinance==0.2.52
streamlit==1.28.0
streamlit-autorefresh==1.0.0
altair==5.1.2
vega-datasets==0.9.0
twilio==8.0.0
fpdf==1.7.2
